feat: refactor template metadata into classes
